IMC calculation done

+++ db/migrate/20170325184754_create_body_masses.rb
@@ -2,9 +2,9 @@ class CreateBodyMasses < ActiveRecord::Migration[5.0]
   def change
     create_table :body_masses do |t|
       t.integer :user_id
-      t.integer :mass
-      t.integer :height
-      t.string :category
+      t.float   :mass
+      t.float   :height
+      t.string  :category
 
       t.timestamps
     end
